+++ src/api/productDetailApi.ts
@@ -1,0 +1,14 @@
+import { Products } from '@/app/lib/products';
+import { createClient } from '../../utils/supabase/server';
+
+export const getProductDetail = async (id: string): Promise<Products> => {
+  const supabase = await createClient();
+
+  const { data, error } = await supabase.from('products').select('*').eq('id', id).single();
+
+  if (error) {
+    throw new Error(error.message);
+  }
+
+  return data;
+};
